+++ web/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./modules/master";
 export * from "./modules/organisation";
+export { coreWebBundle } from "./cxapp";
+export type { CoreWorkspaceContribution } from "./cxapp";